Move duplicate merge code into a utils module
