+++ kasa/iot/modules/usage.py
@@ -4,7 +4,8 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from ..iotmodule import IotModule, merge
+from ...utils import merge
+from ..iotmodule import IotModule
 
 
 class Usage(IotModule):
